feat(workflow-list): unify badge style and remove nested links (#394)
Change the style of the workflow badges (Jupyter notebook, GitHub, ...)
to unify their style. Move the workflow duration to another badge in the
bottom section of the workflow box. Move the actions button to the same
section to avoid having nested interactive element and comply with the
W3 HTML specifications.
