+++ reana-ui/src/pages/workflowList/components/WorkflowBadges.module.scss
@@ -1,0 +1,51 @@
+/*
+  -*- coding: utf-8 -*-
+
+  This file is part of REANA.
+  Copyright (C) 2023 CERN.
+
+  REANA is free software; you can redistribute it and/or modify it
+  under the terms of the MIT License; see LICENSE file for more details.
+*/
+
+.divider {
+  margin: 0 !important;
+  border-color: #f8f8f8 !important;
+}
+
+.badges-and-actions {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  padding: 0.75em 1em;
+
+  .actions {
+    min-width: 22px;
+
+    &:hover {
+      color: #d1d1d1;
+    }
+
+    &.always-visible {
+      visibility: visible;
+    }
+  }
+}
+
+.badgesContainer {
+  display: flex;
+
+  a.label {
+    background-color: transparent;
+    margin-left: 15px;
+
+    &:hover {
+      cursor: pointer;
+      filter: opacity(0.8);
+    }
+
+    &:not(:first-child) {
+      margin-left: 15px;
+    }
+  }
+}
