+++ reana-ui/src/pages/workflowList/components/WorkflowList.js
@@ -8,28 +8,17 @@
   under the terms of the MIT License; see LICENSE file for more details.
 */
 
-import { useSelector } from "react-redux";
-import { Icon, Loader, Message, Popup } from "semantic-ui-react";
+import { Loader, Message } from "semantic-ui-react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
 
-import { INTERACTIVE_SESSION_URL } from "~/client";
-import { getReanaToken } from "~/selectors";
-import {
-  Box,
-  JupyterNotebookIcon,
-  WorkflowActionsPopup,
-  WorkflowDeleteModal,
-  WorkflowStopModal,
-  InteractiveSessionModal,
-} from "~/components";
-import { statusMapping } from "~/util";
+import { Box, WorkflowDeleteModal, WorkflowStopModal, InteractiveSessionModal } from "~/components";
 
 import styles from "./WorkflowList.module.scss";
+import WorkflowDetails from "~/pages/workflowList/components/WorkflowDetails";
+import WorkflowBadges from "~/pages/workflowList/components/WorkflowBadges";
 
 export default function WorkflowList({ workflows, loading }) {
-  const reanaToken = useSelector(getReanaToken);
-
   if (loading) return <Loader active />;
   if (!workflows.length) {
     return <Message info icon="info circle" content="No workflows found." />;
@@ -37,103 +26,15 @@ export default function WorkflowList({ workflows, loading }) {
   return (
     <>
       {workflows.map((workflow) => {
-        const {
-          id,
-          name,
-          run,
-          createdDate,
-          startedDate,
-          finishedDate,
-          friendlyCreated,
-          friendlyStarted,
-          friendlyFinished,
-          duration,
-          completed,
-          total,
-          size,
-          status,
-          session_uri: sessionUri,
-          session_status: sessionStatus,
-        } = workflow;
-        const isDeleted = status === "deleted";
-        const hasDiskUsage = size.raw > 0;
-        const isDeletedUsingWorkspace = isDeleted && hasDiskUsage;
-        const isSessionOpen = sessionStatus === "created";
         return (
-          <Link key={id} to={`/details/${id}`}>
-            <Box
-              className={`${styles.workflow} ${
-                isDeleted ? styles["deleted"] : ""
-              }`}
-            >
-              <div className={styles["details-box"]}>
-                <Icon
-                  className={styles["status-icon"]}
-                  name={statusMapping[status].icon}
-                  color={statusMapping[status].color}
-                />
-                <div>
-                  <span className={styles.name}>{name}</span>
-                  <span className={styles.run}>#{run}</span>
-                  <div>
-                    {hasDiskUsage && (
-                      <span
-                        className={`${styles.size} ${
-                          isDeletedUsingWorkspace ? styles.highlight : ""
-                        }`}
-                      >
-                        <Icon name="hdd" /> {size.human_readable}
-                      </span>
-                    )}
-                    {isSessionOpen && (
-                      <a
-                        href={INTERACTIVE_SESSION_URL(sessionUri, reanaToken)}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        onClick={(e) => e.stopPropagation()}
-                        className={styles.notebook}
-                      >
-                        <JupyterNotebookIcon />
-                      </a>
-                    )}
-                  </div>
-                  <Popup
-                    trigger={
-                      <div>
-                        {friendlyFinished
-                          ? `Finished ${friendlyFinished}`
-                          : friendlyStarted
-                            ? `Started ${friendlyStarted}`
-                            : `Created ${friendlyCreated}`}
-                      </div>
-                    }
-                    content={
-                      friendlyFinished
-                        ? finishedDate
-                        : friendlyStarted
-                          ? startedDate
-                          : createdDate
-                    }
-                  />
-                </div>
+          <Box padding={false} flex={false}>
+            <Link key={workflow.id} to={`/details/${workflow.id}`}>
+              <div className={styles["workflow-details-container"]}>
+                <WorkflowDetails workflow={workflow} actionsOnHover={true} />
               </div>
-              <div className={styles["status-box"]}>
-                <span
-                  className={`${styles["status"]} sui-${statusMapping[status].color}`}
-                >
-                  {status}
-                </span>{" "}
-                {statusMapping[status].preposition} {duration}
-                <div>
-                  step {completed}/{total}
-                </div>
-              </div>
-              <WorkflowActionsPopup
-                workflow={workflow}
-                className={styles.actions}
-              />
-            </Box>
-          </Link>
+            </Link>
+            <WorkflowBadges workflow={workflow} />
+          </Box>
         );
       })}
       <InteractiveSessionModal />
